drivers:i2c: add I2C driver for STM32F4 family
Add a I2C master driver for the STM32F4 family using the LL api.
The driver supports both polling and interrupt.

The master write operation for 7bit addressing was tested on
Carbon from 96Boards.

The master read operation was not tested.
The 10bit address support was not tested.

Slave mode is not implemented.

Origin:Original

Signed-off-by: Yannis Damigos <giannis.damigos@gmail.com>